requirements.txt
fastapi
uvicorn
sqlmodel
httpx
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
email-validator
asyncpg
fastapi-cache2
redis
pydantic-settings
